getGoodAuthorNames
